refactor: refactoring accoring to code-review feedback
