+++ src/colors.scss
@@ -9,3 +9,4 @@ $purple-3: rgba(81, 74, 126, 0.4);
 $gray-1: rgba(125, 125, 125, 1);
 $medium-gray: #7d7d7d;
 $blue: #2465b0;
+$blur: rgba(255, 255, 255, 0.693);
