+++ src/features/main/main.scss
@@ -1,5 +1,4 @@
 main {
-  // height: calc(100% - 50px);
   display: flex;
   flex-direction: column;
   flex: 1 0 auto;
