+++ src/features/slider/slider.scss
@@ -105,7 +105,7 @@ $imageWidth: 575px;
     z-index: 10;
     width: 100%;
     height: 200vh;
-    background: rgba(255, 255, 255, 0.693);
+    background: $blur;
     backdrop-filter: blur(5px);
     opacity: 1;
     transition: all ease-in-out 0.5s;
