+++ src/widgets/product-view/product-view.scss
@@ -1,12 +1,13 @@
+@import '../../colors';
+
 .product-view {
   &__title {
     padding: 60px 0;
-    color: #121212;
+    color: $black-1;
     font-family: Roboto;
-    font-size: 40px;
+    font-size: 34px;
     font-style: normal;
     font-weight: 300;
-    line-height: 87.305%; /* 34.922px */
   }
 
   &__content {
@@ -31,12 +32,11 @@
 
   &__price {
     padding: 10px 0;
-    color: #121212;
+    color: $black-1;
     font-family: Roboto;
-    font-size: 40px;
+    font-size: 34px;
     font-style: normal;
     font-weight: 300;
-    line-height: 87.305%;
 
     &_cross-out {
       text-decoration: line-through;
@@ -56,7 +56,7 @@
   &__description {
     padding: 10px 0;
     max-width: 490px;
-    color: #7d7d7d;
+    color: $medium-gray;
     font-family: Roboto;
     font-size: 16px;
     font-style: normal;
