+++ src/widgets/product-view/product-view.ts
@@ -1,20 +1,19 @@
 import './product-view.scss';
 import { DiscountedPrice, Price, Product } from '@commercetools/platform-sdk';
-import ElementBuilder from '../shared/lib/element-builder';
-import ViewBuilder from '../shared/lib/view-builder';
-import PageTitle from '../features/page-title/page-title';
-import Button from '../shared/ui/button/button';
-import { ButtonSize, ButtonType, ButtonIconPosition } from '../shared/ui/button/models';
-import Slider from '../features/slider/slider';
-import ProductApi from '../entities/product/api';
+import ElementBuilder from '../../shared/lib/element-builder';
+import ViewBuilder from '../../shared/lib/view-builder';
+import PageTitle from '../../features/page-title/page-title';
+import Button from '../../shared/ui/button/button';
+import { ButtonSize, ButtonType, ButtonIconPosition } from '../../shared/ui/button/models';
+import Slider from '../../features/slider/slider';
+import ProductApi from '../../entities/product/api';
 
 export default class ProductView extends ViewBuilder {
-  slides: HTMLElement[];
-  slider: Slider;
-  id: string;
-  data: Product;
-  price: Price;
-  disountedPrice: DiscountedPrice;
+  private slider: Slider;
+  private id: string;
+  private data: Product;
+  private price: Price;
+  private disountedPrice: DiscountedPrice;
 
   constructor() {
     super('product-view');
@@ -40,18 +39,18 @@ export default class ProductView extends ViewBuilder {
     return slides;
   }
 
-  private getPrice(isDiscounted = false) {
+  private getPrice(isDiscounted = false): string {
     this.price = this.data.masterData.current.masterVariant.prices[0];
-    let centAmount = this.price.value.centAmount;
+    let centAmount: number = this.price.value.centAmount;
 
     if (isDiscounted) {
       centAmount = this.price.discounted.value.centAmount;
     }
 
-    const fractionDigits = this.price.value.fractionDigits;
-    const currencyCode = this.price.value.currencyCode;
-    const shortPrice = centAmount / 10 ** fractionDigits;
-    const formatedPrice = new Intl.NumberFormat(`us-US`, {
+    const fractionDigits: number = this.price.value.fractionDigits;
+    const currencyCode: string = this.price.value.currencyCode;
+    const shortPrice: number = centAmount / 10 ** fractionDigits;
+    const formatedPrice: string = new Intl.NumberFormat(`us-US`, {
       style: 'currency',
       currency: `${currencyCode}`,
     }).format(shortPrice);
@@ -59,7 +58,7 @@ export default class ProductView extends ViewBuilder {
     return formatedPrice;
   }
 
-  public configureView() {
+  public configureView(): HTMLElement[] {
     const pageTitle = new PageTitle(this.data.masterData.current.name['en-US']);
     pageTitle.getElement().classList.add('product-view__title');
